extracted tests to separate files
